requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
openai>=1.46.0
# На следующих шагах добавим при необходимости:
# swisseph
# opencv-python
# mediapipe
# reportlab / weasyprint
# requests
